requirements.txt
apache-airflow==3.0.1
requests==2.32.3
duckdb==1.2.2
bs4==0.0.2
pandas == 2.1.4
numpy
lxml